+++ modules/google-analytics/plugin.js
@@ -1,11 +1,11 @@
 // Setup google analytics
-window['GoogleAnalyticsObject'] = 'ga';
+window['GoogleAnalyticsObject'] = 'ga'
 window['ga'] = window['ga'] || function () {
   (window['ga'].q = window['ga'].q || []).push(arguments)
 }
-window['ga'].l = 1 * new Date();
+window['ga'].l = 1 * new Date()
 
-export default ({app: {router}}) => {
+export default ({ app: { router, store } }) => {
   if (!window['ga']) {
     console.warn('google analytics is not available')
     return
@@ -16,8 +16,31 @@ export default ({app: {router}}) => {
 
   // Every time the route changes (fired on initialization too)
   router.afterEach((to, from) => {
-    // We tell Google Analytic to add a page view
+    // Set page settings
+    const settings = Object.assign({}, routeOption('analytics', from, to, store), to.meta && to.meta.analytics)
+    Object.keys(settings).forEach(key => {
+      ga('set', key, settings[key])
+    })
+
+    // We tell Google Analytics to add a page view
     ga('set', 'page', to.fullPath)
     ga('send', 'pageview')
   })
+}
+
+function routeOption(key, from, to, store) {
+  let matched = to.matched[0]
+  let matchedComponent = matched.components.default
+  return componentOption(matchedComponent, key, from, to, store)
+}
+
+function componentOption(component, key, ...args) {
+  if (!component || !component.options || !component.options[key]) {
+    return {}
+  }
+  let option = component.options[key]
+  if (typeof option === 'function') {
+    option = option(...args)
+  }
+  return option
 }