+++ modules/workbox/plugin.js
@@ -1,0 +1,10 @@
+window.onNuxtReady(() => {
+  if (!('serviceWorker' in navigator)) {
+    console.warn('serviceWorker is not supported')
+    return
+  }
+
+  navigator.serviceWorker.register('<%= options.swURL %>', {
+    scope: '<%= options.swScope %>'
+  });
+})