+++ modules/manifest/CHANGELOG.md
@@ -3,6 +3,109 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="1.6.0"></a>
+# [1.6.0](https://github.com/nuxt/modules/compare/@nuxtjs/manifest@1.5.0...@nuxtjs/manifest@1.6.0) (2017-06-06)
+
+
+### Bug Fixes
+
+* **manifest:** correctly omit internal options ([47559b0](https://github.com/nuxt/modules/commit/47559b0))
+
+
+### Features
+
+* **manifest:** meta module compatibility ([8a41fda](https://github.com/nuxt/modules/commit/8a41fda))
+* **manifest:** refactor & openGraph support ([0768246](https://github.com/nuxt/modules/commit/0768246))
+* **manifest:** rewrite module ([352b4bf](https://github.com/nuxt/modules/commit/352b4bf))
+
+
+
+
+<a name="1.5.0"></a>
+# [1.5.0](https://github.com/nuxt/modules/compare/@nuxtjs/manifest@1.4.2...@nuxtjs/manifest@1.5.0) (2017-06-05)
+
+
+### Features
+
+* **manifest:** defaultIcon option ([1086962](https://github.com/nuxt/modules/commit/1086962))
+
+
+
+
+<a name="1.4.2"></a>
+## [1.4.2](https://github.com/nuxt/modules/compare/@nuxtjs/manifest@1.4.1...@nuxtjs/manifest@1.4.2) (2017-06-04)
+
+
+### Bug Fixes
+
+* prevent invalid url when router base is / ([f0fd863](https://github.com/nuxt/modules/commit/f0fd863))
+
+
+
+
+<a name="1.4.1"></a>
+## [1.4.1](https://github.com/nuxt/modules/compare/@nuxtjs/manifest@1.4.0...@nuxtjs/manifest@1.4.1) (2017-06-04)
+
+
+### Bug Fixes
+
+* **manifest:** sanetize options ([b8497a0](https://github.com/nuxt/modules/commit/b8497a0))
+* **manifest:** typo in manifest.json filename ([c2cabb6](https://github.com/nuxt/modules/commit/c2cabb6))
+
+
+
+
+<a name="1.4.0"></a>
+# [1.4.0](https://github.com/nuxt/modules/compare/@nuxtjs/manifest@1.3.1...@nuxtjs/manifest@1.4.0) (2017-06-04)
+
+
+### Features
+
+* **manifest:** improvements ([cac9b4e](https://github.com/nuxt/modules/commit/cac9b4e))
+
+
+
+
+<a name="1.3.1"></a>
+## [1.3.1](https://github.com/nuxt/modules/compare/@nuxtjs/manifest@1.3.0...@nuxtjs/manifest@1.3.1) (2017-06-02)
+
+
+### Bug Fixes
+
+* **manifest:** write manifest file once ([18aa015](https://github.com/nuxt/modules/commit/18aa015))
+
+
+
+
+<a name="1.3.0"></a>
+# [1.3.0](https://github.com/nuxt/modules/compare/@nuxtjs/manifest@1.2.0...@nuxtjs/manifest@1.3.0) (2017-06-02)
+
+
+### Features
+
+* **manifest:** add lang ([bfdb96e](https://github.com/nuxt/modules/commit/bfdb96e))
+* **manifest:** improve default short_name ([45b2bb2](https://github.com/nuxt/modules/commit/45b2bb2))
+
+
+
+
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/nuxt/modules/compare/@nuxtjs/manifest@1.1.1...@nuxtjs/manifest@1.2.0) (2017-06-02)
+
+
+### Features
+
+* **meta:** add apple-touch-icon ([c74ffa4](https://github.com/nuxt/modules/commit/c74ffa4))
+
+
+
+
+<a name="1.1.1"></a>
+## [1.1.1](https://github.com/nuxt/modules/compare/@nuxtjs/manifest@1.1.0...@nuxtjs/manifest@1.1.1) (2017-05-31)
+
+
+
+
 <a name="1.1.0"></a>
 # [1.1.0](https://github.com/nuxt/modules/compare/@nuxtjs/manifest@1.0.1...@nuxtjs/manifest@1.1.0) (2017-05-30)