+++ modules/meta/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="1.3.0"></a>
+# [1.3.0](https://github.com/nuxt/modules/compare/@nuxtjs/meta@1.2.1...@nuxtjs/meta@1.3.0) (2017-07-20)
+
+
+### Features
+
+* **icon:** applefavicon option (#76) ([280b416](https://github.com/nuxt/modules/commit/280b416))
+
+
+
+
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/nuxt/modules/compare/@nuxtjs/meta@1.1.0...@nuxtjs/meta@1.2.0) (2017-06-06)
+
+
+### Features
+
+* **meta:** Support manifest meta & openGraph ([17b7db1](https://github.com/nuxt/modules/commit/17b7db1))
+
+
+
+
+<a name="1.1.0"></a>
+# [1.1.0](https://github.com/nuxt/modules/compare/@nuxtjs/meta@1.0.1...@nuxtjs/meta@1.1.0) (2017-06-02)
+
+
+### Features
+
+* **meta:** more meta tags ([a4e3a04](https://github.com/nuxt/modules/commit/a4e3a04))
+
+
+
+
 <a name="1.0.1"></a>
 ## [1.0.1](https://github.com/nuxt/modules/compare/@nuxtjs/meta@1.0.0...@nuxtjs/meta@1.0.1) (2017-05-29)